Done pause, can resume, go back to main menu page properly, bullet collides properly with enemy

+++ Assignment1/app/src/main/java/sidm/com/assignment1/Enemy.java
@@ -5,10 +5,10 @@ public abstract class Enemy extends GameObject
     protected Health health = new Health();
     protected Weapon weapon = null;
     protected boolean isDead = false;
-    protected float dmg = 0f;
-    protected float score = 0f;
+    protected int dmg = 0;
+    protected int score = 0;
 
-    public void AddHealth(final float amt)
+    public void AddHealth(final int amt)
     {
         health.AddHealth(amt);
         if(health.GetHealth() <= 0.f)
@@ -18,15 +18,15 @@ public abstract class Enemy extends GameObject
     // Getter
     final public boolean GetIsDead() { return isDead; }
     final public Weapon GetWeapon() { return weapon; }
-    final public float GetDamage() { return dmg; }
+    final public int GetDamage() { return dmg; }
     final public Health GetHealth() { return health; }
-    final public float GetScore() { return score; }
+    final public int GetScore() { return score; }
 
     // Setter
     public void SetIsDead(final boolean _isDead) { isDead = _isDead; }
     public void SetWeapon(final Weapon _weapon) { weapon = _weapon; }
-    public void SetDamage(final float _dmg) { dmg = _dmg; }
-    public void SetScore(final float _score) { score = _score; }
+    public void SetDamage(final int _dmg) { dmg = _dmg; }
+    public void SetScore(final int _score) { score = _score; }
 
     @Override
     public int GetRenderLayer() { return LayerConstants.GAMEOBJECTS_LAYER; }
